refactor: YearSemester Enum를 common 패키지로 이동
